+++ api/products.js
@@ -419,6 +419,19 @@ async function deleteProduct(req, res) {
     return res.status(404).json({ error: 'Product not found' });
   }
 
+  // Athuga hvort vara sé til í körfu/pöntun
+  const countQuery = 'SELECT COUNT(*) FROM orderLines WHERE product_id = $1';
+  const countResult = await query(countQuery, [id]);
+
+  const { count } = countResult.rows[0];
+
+  // Leyfum bara að eyða tómum flokkum
+  if (toPositiveNumberOrDefault(count, 0) > 0) {
+    return res.status(400).json({
+      error: 'Product exists in cart or order, cannot delete.',
+    });
+  }
+
   const q = 'DELETE FROM products WHERE ID = $1';
   await query(q, [id]);
